+++ CHANGELOG.rst
@@ -20,7 +20,7 @@ before applying to production despite of the absence of warning marks.
 4.1.0-stable
 ============
 
-* TBA: Video recording server. Bookmarks in vendored apps.
+* TBA: Video recording server. Bookmarks in vendored apps. svc.flash core method.
 
 4.0.2-stable
 ============